--- VitalLog.cs
@@ -1,10 +1,0 @@
-using System;
-using System.Diagnostics;
-
-class VitalLog
-{
-    public static void getVitalLog(string msg)
-    {
-        Console.WriteLine(msg);
-    }
-}
